requirements.txt
flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.4
requests==2.31.0
roboflow==1.1.1
python-dotenv==1.0.0